json de-/serialisation from/to string or file (#1207)
* move unique_ptr and cleanup unused

* commit before updating to master

* remove warning

* ak_from_json and tests added

* test added

* pre-commit formatting

* add a test

* add tests

* Leave 'pybind11' at version 2.9.0.

* use the same functions for both versions

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* rebase and fix an error reported by pre commit

* add tests to read from json files

* cleanup

* to json and tests

* pass parameters

* to_json without special encoder class

* rebase and pre-commit

* add test for complex with nan and inf

* cleanup

* disable pylint's unspecified-encoding when opening a binary file

* read file as text and add encoding

* change encoding to utf-8

* add encoding

* drop RapidJson for from_json, first go at ndjson:
- accept newline as line delimiter '\n' (0x0A) as well as carriage return and newline '\r\n' (0x0D0A)
- ignore empty lines '\n\n'
- raise an error if JSON is nor parsable

* cleanup

* remove dependency on rapidjson, implement ndjson specs

* user-defined strings postprocessing

* final cleanup

* all done

+++ src/libawkward/builder/DatetimeBuilder.cpp
@@ -42,7 +42,7 @@ namespace awkward {
 
     container.copy_buffer(form_key.str() + "-data",
                           content_.ptr().get(),
-                          content_.length() * (int64_t)sizeof(int64_t));
+                          (int64_t)(content_.length() * sizeof(int64_t)));
 
     std::string primitive(units_);
 
@@ -67,7 +67,7 @@ namespace awkward {
 
   int64_t
   DatetimeBuilder::length() const {
-    return content_.length();
+    return (int64_t)content_.length();
   }
 
   void
